+++ app/controllers/redmine_oauth_controller.rb
@@ -43,6 +43,12 @@ class RedmineOauthController < AccountController
         state: oauth_csrf_token,
         scope: 'read_user'
       )
+    when 'Google'
+      redirect_to oauth_client.auth_code.authorize_url(
+        redirect_uri: oauth_callback_url,
+        state: oauth_csrf_token,
+        scope: 'openid:profile:email'
+      )
     when 'Okta'
       redirect_to oauth_client.auth_code.authorize_url(
         redirect_uri: oauth_callback_url,
@@ -71,6 +77,11 @@ class RedmineOauthController < AccountController
       token = oauth_client.auth_code.get_token(params['code'], redirect_uri: oauth_callback_url)
       userinfo_response = token.get('/api/v4/user', headers: { 'Accept' => 'application/json' })
       user_info = JSON.parse(userinfo_response.body)
+      user_info['login'] = user_info['username']
+      email = user_info['email']
+    when 'Google'
+      token = oauth_client.auth_code.get_token(params['code'], redirect_uri: oauth_callback_url)
+      user_info = JWT.decode(token.token, nil, false).first
       user_info['login'] = user_info['username']
       email = user_info['email']
     when 'Okta'
@@ -169,6 +180,14 @@ class RedmineOauthController < AccountController
           site: site,
           authorize_url: '/oauth/authorize',
           token_url: '/oauth/token'
+        )
+      when 'Google'
+        OAuth2::Client.new(
+          Setting.plugin_redmine_oauth[:client_id],
+          Setting.plugin_redmine_oauth[:client_secret],
+          site: site,
+          authorize_url: '/o/oauth2/v2/auth',
+          token_url: '/o/oauth2/v2/auth'
         )
       when 'Okta'
         OAuth2::Client.new(