+++ assets/javascripts/redmine_oauth.js
@@ -70,6 +70,11 @@ function oauth_settings_visibility()
             div_oauth_options.show();
             div_oauth_options.find('#oauth_options_tenant').hide();
             break;
+        case 'Keycloak':
+            div_oauth_options.show();
+            div_oauth_options.find('#oauth_options_tenant').show();
+            tenant_id.val("");
+            break;
         case 'Okta':
             div_oauth_options.show();
             div_oauth_options.find('#oauth_options_tenant').show();