+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for Redmine OAuth
     
     Autologin
     Google OAuth provider
+    Keycloak OAuth provider
 
 2.1.3 *2023-11-20*
 ------------------